requirements.txt
# Core dependencies
pytest>=7.0.0

# Development dependencies
pytest-cov>=4.0.0

# LLM integration
litellm>=1.0.0

# Configuration management
pyyaml>=6.0.0

# CLI enhancements
click>=8.0.0
rich>=13.0.0

# Future dependencies (for later phases)
# fastapi>=0.100.0
# uvicorn>=0.20.0